pyproject.toml
[project]
name = "agentic-blog-writing-system-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[uvicorn]>=0.128.2",
    "langchain>=1.2.8",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.7",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
